Added a prompt to destroy when the user supplies a name that is different than what was generated. Using filter instead of splice to remove the reducer from index.js

+++ src/commands/destroy.js
@@ -1,6 +1,7 @@
 var fs = require("fs");
 var path = require("path");
 var chalk = require("chalk");
+var inquirer = require("inquirer");
 
 var availableCommands = {
   "container": "containers",
@@ -8,7 +9,8 @@ var availableCommands = {
   "reducer": "reducers"
 };
 
-module.exports = function () {
+// @NOTE, we use async here because we use await later in the function
+module.exports = async function () {
   var command = process.argv[3];
   var name = process.argv[4];
 
@@ -41,6 +43,7 @@ module.exports = function () {
   }
 
   // Step 4: Possibly mutate the name by converting it to Pascal Case (only for container and component for now)
+  var originalName = name; // store original name for later
   if (["container", "component"].indexOf(command) !== -1) {
     name = name.substr(0, 1).toUpperCase() + name.substr(1);
   }
@@ -59,8 +62,27 @@ module.exports = function () {
   }
 
   if (fileExists) {
+    // If they typed a name that would normally be mutated, check with the user first before deleting the files
+    if (originalName !== name) {
+      // @NOTE: We are using await so that we can wait for the result of the promise before moving on
+      const continueDestroying = await new Promise((resolve, reject) => {
+        const question = {
+          type: "confirm",
+          name: "confirm",
+          message: `You wanted to destroy ${chalk.yellow(originalName)} but the generated name is ${chalk.green(name)}. Would you like to continue with destroying ${chalk.green(name)}?`
+        };
+        inquirer.prompt([question], function (answers) {
+          if (!answers.confirm) resolve(false);
+          resolve(true);
+        });
+      });
+
+      // Since we used await, this code will not be executed until the promise above resolves
+      if (!continueDestroying) process.exit();
+    }
+
     fs.unlinkSync(destinationPath);
-    console.log(chalk.red(`Removed file: ${destinationPath}`));
+    console.log(`${chalk.red("Removed file:")} ${destinationPath}`);
   }
   else {
     console.log(chalk.red(`ERROR: ${chalk.yellow(destinationPath)} does not exist`));
@@ -71,12 +93,15 @@ module.exports = function () {
   if (command === "reducer") {
     var reducerIndexPath = path.resolve(process.cwd(), "src/reducers/index.js");
     try {
-      var indexLines = fs.readFileSync(reducerIndexPath, {encoding: "utf8"}).split("\n");
-      var reducerLine = `export { default as ${name} } from "./${name}"`;
-      var reducerLineIndex = indexLines.indexOf(reducerLine);
-      indexLines.splice(reducerLineIndex, 1);
-      fs.writeFileSync(reducerIndexPath, indexLines.join("\n"));
-      console.log(chalk.red(`${name} removed from reducer index ${reducerIndexPath}`));
+      const indexLines = fs.readFileSync(reducerIndexPath, {encoding: "utf8"}).split("\n");
+      const reducerLine = `export { default as ${name} } from "./${name}"`;
+      const newIndexLines = indexLines.filter((indexLine) => {
+        // Only return lines from the reducer index that do not return the reducer
+        // we just destroyed. Check for semi colon and non-semi colon lines
+        return indexLine !== reducerLine && indexLine !== `${reducerLine};`;
+      });
+      fs.writeFileSync(reducerIndexPath, newIndexLines.join("\n"));
+      console.log(chalk.red(`${name} removed`) + ` from reducer index ${reducerIndexPath}`);
     }
     catch (e) {
       console.log(chalk.red(`ERROR: Unable to modify reducers index. Reducer not removed from index`));
@@ -85,7 +110,7 @@ module.exports = function () {
   }
 
   // Step 7: Remove the test file for a component
-  if(command === "component") {
+  if (command === "component") {
     var testPath = path.join(process.cwd(), "/test/", availableCommands[command], `/${name}.test.js`);
     var testFileExists = true;
     try {
